refactor: consolidate types and dependencies
