+++ apps/web/hooks/index.ts
@@ -1,2 +1,2 @@
 export { useFlags } from './useFlags';
-export { useWebEngine } from './useWebEngine';
+export { useComponentMetrics } from './useComponentMetrics';
