+++ apps/web/package.json
@@ -13,9 +13,7 @@
   },
   "dependencies": {
     "@bos-web-engine/application": "workspace:*",
-    "@bos-web-engine/compiler": "workspace:*",
-    "@bos-web-engine/container": "workspace:*",
-    "@bos-web-engine/iframe": "workspace:*",
+    "@bos-web-engine/common": "workspace:*",
     "next": "^13.5.6",
     "react": "18.2.0",
     "react-dom": "18.2.0",
