+++ packages/common/src/types/render.ts
@@ -1,0 +1,18 @@
+export interface WebEngineMeta {
+  componentId?: string;
+  isProxy?: boolean;
+  parentMeta?: WebEngineMeta;
+}
+
+export interface KeyValuePair {
+  [key: string]: any;
+}
+
+export interface Props extends KeyValuePair {
+  __bweMeta?: WebEngineMeta;
+  __domcallbacks?: { [key: string]: any };
+  __componentcallbacks?: { [key: string]: any };
+  children?: any[];
+  className?: string;
+  id?: string;
+}
